backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
